Address PR Comment
